fix(#44): move operation_types.py + rpc_types.py from core/ to contracts/
Pure data types (StrEnum, dataclasses) with no kernel dependencies.
- Move core/operation_types.py → contracts/operation_types.py (2 callers)
- Move core/rpc_types.py → contracts/rpc_types.py (4 callers)
- Update all 6 import sites
