+++ src/nexus/contracts/operation_types.py
