+++ src/nexus/server/protocol.py
@@ -14,10 +14,9 @@ from __future__ import annotations
 
 from typing import Any
 
-# Issue #1519, 1A: RPC types extracted to core/rpc_types.py so core/ modules
-# (rpc_transport, rpc_codec) can use them without importing from server/.
+# Issue #1519, 1A: RPC types live in contracts/rpc_types.py (pure data types).
 # Re-exported here for backward compatibility.
-from nexus.core.rpc_types import RPCErrorCode, RPCRequest, RPCResponse  # noqa: F401
+from nexus.contracts.rpc_types import RPCErrorCode, RPCRequest, RPCResponse  # noqa: F401
 
 # ============================================================
 # Generated Param classes + METHOD_PARAMS
