+++ src/nexus/services/rpc_transport.py
@@ -121,8 +121,8 @@ class NexusRPCTransport:
             TransportError: Connection or timeout error
             RPCError: Server returned an error response
         """
-        # Import codec and protocol types from core/ (Issue #1519, 1A)
-        from nexus.core.rpc_types import RPCErrorCode, RPCRequest, RPCResponse
+        # Import codec and protocol types from contracts/ (Issue #1519, 1A)
+        from nexus.contracts.rpc_types import RPCErrorCode, RPCRequest, RPCResponse
         from nexus.lib.rpc_codec import decode_rpc_message, encode_rpc_message
 
         request = RPCRequest(
